create ContentSection component, add public media, update layout
